Updating the CSS to match CSS in-app
The CSS in these style sheets have not been updated to match the styling
added to the css style sheets in-app. These changes include adding more
control over width and padding, adjusting margins, and so on. This will
help to provide customers with accurate representations of the themes
we provide in-app.